requirements.txt
Django==4.2.0
Pillow==10.4.0
python-dateutil==2.9.0
requests==2.32.3
